refactor(server): split httpCommon, unify rate-limit and auth lookup (wave 2/5)
Хвиля 2 із 5 плану прибирання дублювання бекенду.

http-інфраструктура:
- Розпакувати `server/httpCommon.mjs` (287 LoC, 5 доменів в одному файлі) на
  `server/http/{middleware,security,authMiddleware,health,errorHandler}.js`
  + barrel `server/http/index.js`.
- Старий `httpCommon.mjs` видалено; `server/app.js` імпортує з нового barrel.
- CSP-тести перенесено у `server/http/security.test.js`.

rate-limit:
- Прибрати дублікат `checkRateLimit` у
  `server/api/nutrition/lib/nutritionSecurity.js` (своя in-memory `buckets`
  Map — два незалежні bucket-и для того самого IP).
- Єдине джерело — `server/api/lib/rateLimit.js`. Додано `retryAfterSec` у
  публічний контракт `checkRateLimit` (уже використовується трьома nutrition-
  хендлерами) і підʼєднано до `Retry-After` у `rateLimitExpress`.
- `nutritionSecurity.js` лишає `requireNutritionTokenIfConfigured` +
  re-export `checkRateLimit` для зворотної сумісності існуючих імпортів.

auth:
- `server/api/push.js::getSession` дублював `auth.js::getSessionUser`, але
  без запису метрик/ALS-контексту. Замінено на виклик `getSessionUser`,
  обгорнутий у try/catch (push історично повертає 401 на будь-яку помилку
  better-auth).
- `server/auth.js::getSessionUser` перевів на `fromNodeHeaders` з
  `better-auth/node` (більш надійна конвертація, ніж `new Headers`).

Контракт API без змін. Тести: 669 pass (71 files), lint/typecheck зелені.
Smoke: `/livez` 200, `/api/push/vapid-public` 503 (expected без VAPID),
OPTIONS /api/nutrition/analyze-photo 200, GET → 405.
