+++ server/http/index.js
@@ -1,0 +1,25 @@
+/**
+ * Barrel-module — єдина точка імпорту HTTP-інфраструктури для `server/app.js`
+ * та тестів. Реалізації розбито по доменних файлах у цій теці; якщо треба
+ * додати новий middleware — клади його у відповідний файл і ре-експортуй тут.
+ */
+export {
+  requestIdMiddleware,
+  requestLogMiddleware,
+  withRequestContext,
+} from "./middleware.js";
+
+export { buildApiCspDirectives, apiHelmetMiddleware } from "./security.js";
+
+export {
+  authSensitiveRateLimit,
+  authMetricsMiddleware,
+} from "./authMiddleware.js";
+
+export {
+  livezHandler,
+  createReadyzHandler,
+  createHealthHandler,
+} from "./health.js";
+
+export { errorHandler } from "./errorHandler.js";
